+++ notifier/main.go
@@ -1,0 +1,8 @@
+package notifier
+
+// ----------------------------------------------------------------------------
+// Constants
+// ----------------------------------------------------------------------------
+
+// Identfier of the  package found messages having the format "senzing-6461xxxx".
+const ProductId = 6461